requirements.txt
altgraph==0.17
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
future==0.18.2
geographiclib==1.50
geopy==2.1.0
idna==2.10
isodate==0.6.0
lxml==4.6.2
numpy==1.19.4
opencv-python==4.4.0.46
pefile==2019.4.18
pycep-correios==4.0.4
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyTelegramBotAPI==3.7.4
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
pywin32-ctypes==0.2.0
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
validate-docbr==1.8.2
wrapt==1.12.1
zeep==4.0.0
